backend/requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
dataclasses==0.8
Django==3.0.7
django-filter==2.3.0
django-rest-knox==4.0.1
djangorestframework==3.11.0
djongo==1.3.6
gunicorn==20.0.4
importlib-metadata==4.5.0
Markdown==3.2.2
mysqlclient==2.0.3
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.6
pycparser==2.20
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
typing-extensions==3.10.0.0
zipp==3.4.1
